Avoid using the contained language system for inteceptors so we can more easily change serialization later
